requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.38.0
requests>=2.32.0
python-dotenv>=1.0.1
pydantic-settings>=2.5.0
tiktoken>=0.8.0
google-api-python-client>=2.140.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1

llama-index>=0.11.0
llama-index-vector-stores-qdrant>=0.3.0
llama-index-readers-file>=0.2.0
llama-index-readers-json>=0.2.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0
qdrant-client>=1.11.0

pypdf>=5.0.0
